+++ src/test/jtx/import.h
@@ -1,7 +1,7 @@
 //------------------------------------------------------------------------------
 /*
     This file is part of rippled: https://github.com/ripple/rippled
-    Copyright (c) 2019 Ripple Labs Inc.
+    Copyright (c) 2023 XRPL Labs
 
     Permission to use, copy, modify, and/or distribute this software for any
     purpose  with  or without fee is hereby granted, provided that the above
@@ -60,4 +60,4 @@ loadXpop(std::string content);
 }  // namespace test
 }  // namespace ripple
 
-#endif
+#endif // RIPPLE_TEST_JTX_IMPORT_H_INCLUDED
