+++ src/test/rpc/LedgerClosed_test.cpp
@@ -35,7 +35,6 @@ public:
         env.fund(XRP(10000), alice);
 
         auto lc_result = env.rpc("ledger_closed")[jss::result];
-        std::cout << lc_result[jss::ledger_hash] << "\n";
         BEAST_EXPECT(
             lc_result[jss::ledger_hash] ==
             "2E6983FE0D18F04568199D75E8E69503E747A24CE8814CF1FAD638CF6DF8FBDC");
@@ -51,8 +50,6 @@ public:
         BEAST_EXPECT((*ar_alice)[sfBalance] == XRP(10000));
 
         lc_result = env.rpc("ledger_closed")[jss::result];
-        std::cout << "lc_result[jss::ledger_hash]: "
-                  << lc_result[jss::ledger_hash] << "\n";
         BEAST_EXPECT(
             lc_result[jss::ledger_hash] ==
             "D39C52DE7CBF561ECA875A6D636B7C9095408DE1FAF4EC4AAF3FDD8AB3A1EA55");
